+++ src/mon-move.h
@@ -26,6 +26,7 @@ enum monster_stagger {
 };
 
 bool multiply_monster(const struct monster *mon);
+bool monster_find_enemy(struct monster *mon);
 void process_monsters(int minimum_energy);
 void reset_monsters(void);
 void restore_monsters(void);